Modularized the sub components of the form.
