+++ client/src/components/Form/components/index.js
@@ -1,0 +1,3 @@
+export { FormField } from "./FormField"
+export { FormInput } from "./FormInput"
+export { RangeInput } from "./RangeInput"
